Fail the Github Actions step when the command errors (#21)
With the `markdown-extract` command wrapped in a subshell, if it fails,
the entrypoint script itself doesn't produce an error. Instead, the
`Error: No matches.` output of the command (which SHOULD go to stderr,
not stdout) ends up in the `markdown` output and the entire action step
succeeds (when it SHOULD fail).